requirements.txt
faiss-cpu==1.9.0.post1
fastapi==0.115.6
langchain-core==0.3.29
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
mysql-connector-python==9.1.0
numpy==1.26.4
pydantic==2.10.5
pypdf==5.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
